requirements.txt
torch>=1.9.0
transformers>=4.15.0
huggingface-hub>=0.10.0
datasets>=2.0.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
pyyaml>=6.0
mlflow>=1.20.0
pytest>=6.0.0
tqdm>=4.62.0
matplotlib>=3.5.0
seaborn>=0.11.0